Allow custom error messages.
When an error occurs, the server sends an error code and the JavaScript
has a list of translated error messages for each error code.
When unkown error code is sent, previously it wasn't displayed anything.
By changing the behavior to displaying the error code as fallback it is
possible to customize the server part and send a custom error
message (as error code), which will be displayed directly.