Use assertIn/notIn instead of deprecated methods.
Also one test was incorrect, in the uninstall test 'QuickUpload' shouldn't be
in the installed list. The test seems to have started failing at some point, but
i wonder why it passed before. Anyway, now the tests should be correct.

Skip tests for plone 4.2 as they do not seem to work. This is completely
unrelated to my changes though.